visualisation/backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
cachetools
pydantic-settings
python-dotenv
pandas
pyarrow
db-dtypes
